-Изменён код импорта excel-файла в систему. Убрана обязательная проверка на соответствие по какой-либо из колонок; -Так же вышеописанная проверка была убрана из контроллеров; -Обновление модели (т.к. тип старого значения Price был изменён с long на decimal);
git-tfs-id: [http://192.168.3.86:8080/tfs/defaultcollection]$/CIPJS/KadOzenka;C56613

+++ Source/KadOzenka.Web/Models/MarketObject/CoreObjectDto.cs
@@ -26,7 +26,7 @@ namespace KadOzenka.Web.Models.MarketObject
 		public string Metro { get; set; }
 		public decimal? Area { get; set; }
 		public string Description { get; set; }
-		public long? Price { get; set; }
+		public decimal? Price { get; set; }
 		public string CadastralNumber { get; set; }
 		public string CadastralQuartal { get; set; }
 		public string District { get; set; }
